25/12 Update 3
Aparentemente funcional en local

+++ microservicios/backend/client-service/src/main/java/com/tutorial/client_service/service/ClientService.java
@@ -5,6 +5,9 @@ import com.tutorial.client_service.model.Loan;
 import com.tutorial.client_service.model.LoanData;
 import com.tutorial.client_service.repository.ClientRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 
@@ -19,6 +22,18 @@ public class ClientService {
 
     @Autowired
     RestTemplate restTemplate;
+
+    @Autowired
+    public void setRestTemplate(RestTemplate restTemplate) {
+        restTemplate.getInterceptors().add((request, body, execution) -> {
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            if (authentication instanceof JwtAuthenticationToken jwtAuthenticationToken) {
+                request.getHeaders().setBearerAuth(jwtAuthenticationToken.getToken().getTokenValue());
+            }
+            return execution.execute(request, body);
+        });
+        this.restTemplate = restTemplate;
+    }
 
     public List<Client> getAll() {
         return clientRepository.findAll();
@@ -38,8 +53,13 @@ public class ClientService {
     }
 
     public boolean delete(Long id) {
-        clientRepository.deleteById(id);
-        return true;
+        Client client = getClientById(id);
+        if (client.getLoans().isEmpty()) {
+            clientRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 
     public List<Client> getClientsWithDelayedLoans(){
